Sale un error cuando se intenta eliminar un vehiculo con reservas

+++ src/main/java/Projecte2/DriftoCar/service/MySQL/VehicleService.java
@@ -7,6 +7,7 @@ package Projecte2.DriftoCar.service.MySQL;
 import Projecte2.DriftoCar.entity.MySQL.Localitzacio;
 import Projecte2.DriftoCar.entity.MySQL.Vehicle;
 import Projecte2.DriftoCar.repository.MySQL.LocalitzacioRepository;
+import Projecte2.DriftoCar.repository.MySQL.ReservaRepository;
 import Projecte2.DriftoCar.repository.MySQL.VehicleRepository;
 import java.time.LocalDate;
 import java.util.List;
@@ -32,6 +33,9 @@ public class VehicleService {
     @Autowired
     private LocalitzacioRepository localitzacioRepository;
 
+    @Autowired
+    private ReservaRepository reservaRepository;
+
     public Vehicle altaVehicle(Vehicle vehicle) {
         log.info("S'ha entrat al metode d'altaVehicle.");
 
@@ -45,6 +49,11 @@ public class VehicleService {
         log.info("S'ha entrat al metode de baixaReserva.");
         if (!vehicleRepository.existsById(matricula)) {
             throw new RuntimeException("No existeix cap vehicle amb la matrícula: " + matricula);
+        }
+        boolean tieneReservas = reservaRepository.existsByVehicleMatricula(matricula);
+
+        if (tieneReservas) {
+            throw new RuntimeException("No es pot esborrar un vehicle amb reserves actives.");
         }
         vehicleRepository.deleteById(matricula);
     }
